testing mice option enabled

+++ code/mice.py
@@ -15,7 +15,7 @@ import itertools
 # custom 
 import plotter
 import datahandler as dh
-from cpp_interface import evaluate_expert, self_play
+from cpp_interface import evaluate_expert, test_evaluate_expert, self_play
 from param import Param 
 # from learning.discrete_emptynet import DiscreteEmptyNet
 from learning.continuous_emptynet import ContinuousEmptyNet
@@ -30,7 +30,7 @@ def my_loss(value, policy, target_value, target_policy, weight, z_mu, z_logvar):
 
 	value_loss = weighted_mse_loss(value,target_value,weight)
 	policy_recon_loss = weighted_mse_loss(policy,target_policy,weight.unsqueeze(1))
-	policy_kl_loss = torch.sum(0.5 * torch.sum(torch.exp(z_logvar) + torch.square(z_mu) - (1. + z_logvar), axis=1),axis=0)
+	policy_kl_loss = (1/value.shape[0]) * torch.sum(0.5 * torch.sum(torch.exp(z_logvar) + torch.square(z_mu) - (1. + z_logvar), axis=1),axis=0)
 	policy_loss = policy_recon_loss + policy_kl_loss
 
 	return value_loss + policy_loss
@@ -292,47 +292,10 @@ def load_param(some_dict):
 	return param 
 
 def evaluate_expert_wrapper(arg):
+	evaluate_expert(*arg)
+
+def test_evaluate_expert_wrapper(arg):
 	test_evaluate_expert(*arg)
-	# evaluate_expert(*arg)
-
-def test_evaluate_expert(states,param,quiet_on=True,progress=None):
-	
-	if not quiet_on:
-		print('   running expert for instance {}'.format(param.dataset_fn))
-
-	if progress is not None:
-		progress_pos = current_process()._identity[0] - progress
-		enumeration = tqdm.tqdm(states, desc=param.dataset_fn, leave=False, position=progress_pos)
-	else:
-		enumeration = states
-
-	game = param_to_cpp_game(param)
-
-	sim_result = {
-		'states' : [],
-		'policy_dists' : [],
-		'values' : [],
-		'param' : param.to_dict()
-		}
-
-	for state in enumeration:
-		game_state = state_to_cpp_game_state(param,state,param.training_team)
-		mctsresult = mctscpp.search(game, game_state,
-			param.mcts_tree_size, param.mcts_rollout_beta, param.mcts_c_param,
-			param.mcts_pw_C, param.mcts_pw_alpha)
-		if mctsresult.success: 
-			policy_dist = valuePerAction_to_policy_dist(param,mctsresult.valuePerAction) # 
-			value = mctsresult.expectedReward[0]
-			sim_result["states"].append(state) # total number of robots x state dimension per robot 
-			sim_result["policy_dists"].append(policy_dist)  
-			sim_result["values"].append(value)
-
-	sim_result["states"] = np.array(sim_result["states"])
-	sim_result["policy_dists"] = np.array(sim_result["policy_dists"])
-	dh.write_sim_result(sim_result,param.dataset_fn)
-
-	if not quiet_on:
-		print('   completed instance {} with {} dp.'.format(param.dataset_fn,sim_result["states"].shape[0]))	
 
 def make_dataset(states,params,df_param):
 	print('making dataset...')
@@ -351,7 +314,10 @@ def make_dataset(states,params,df_param):
 		num_workers = min(ncpu-1, len(params))
 		with Pool(num_workers, initializer=tqdm.set_lock, initargs=(tqdm.get_lock(),)) as p:
 			args = list(zip(states, params, itertools.repeat(True), itertools.repeat(pool_count)))
-			r = list(tqdm(p.imap_unordered(evaluate_expert_wrapper, args), total=len(args), position=0))
+			if df_param.mice_testing_on:
+				r = list(tqdm(p.imap_unordered(test_evaluate_expert_wrapper, args), total=len(args), position=0))
+			else:
+				r = list(tqdm(p.imap_unordered(evaluate_expert_wrapper, args), total=len(args), position=0))
 			# p.starmap(evaluate_expert, states, params)
 			# p.starmap(evaluate_expert, list(zip(states, params)))
 		pool_count += num_workers
@@ -427,12 +393,40 @@ def format_dir(df_param):
 				os.remove(file)
 		os.makedirs(datadir,exist_ok=True)
 
-	modeldir = df_param.path_current_models
-	if os.path.exists(modeldir):
-		for file in glob.glob(modeldir + "/*"):
-			os.remove(file)
-	os.makedirs(modeldir,exist_ok=True)	
+	if df_param.clean_models_on:
+		modeldir = df_param.path_current_models
+		if os.path.exists(modeldir):
+			for file in glob.glob(modeldir + "/*"):
+				os.remove(file)
+		os.makedirs(modeldir,exist_ok=True)	
 
+def test_model(param,model_fn):
+
+	stats = {
+		'values' : [],
+		'policies' : [],
+	}
+
+	test_state = np.array([
+		[0.1, 0.1, 0.0, 0.0],
+		[0.3, 0.3, 0.0, 0.0]
+		])
+
+	robot_idx = 0 
+	n_samples = 100 
+
+	model = ContinuousEmptyNet(param, "cpu")
+	model.load_state_dict(torch.load(model_fn))
+
+	o_a,o_b,goal = relative_state(test_state,param,robot_idx)
+	o_a,o_b,goal = format_data(o_a,o_b,goal)
+
+	for i_sample in range(n_samples):
+		value, policy = model(o_a,o_b,goal)
+		stats["values"].append(value)
+		stats["policies"].append(policy)
+
+	return stats
 
 if __name__ == '__main__':
 
@@ -440,7 +434,9 @@ if __name__ == '__main__':
 
 	df_param = Param() 
 	df_param.clean_data_on = True
+	df_param.clean_models_on = True
 	df_param.make_data_on = True
+	df_param.mice_testing_on = False
 
 	print('Clean old data on: {}'.format(df_param.clean_data_on))
 	print('Make new data on: {}'.format(df_param.make_data_on))
@@ -448,7 +444,7 @@ if __name__ == '__main__':
 	format_dir(df_param)
 
 	# create randomly initialized models for use in the first iteration
-	model = ContinuousEmptyNet(df_param,'cpu')
+	model = ContinuousEmptyNet(df_param,df_param.device)
 	for training_team in df_param.l_training_teams:
 		torch.save(model.state_dict(), df_param.l_model_fn.format(\
 						DATADIR=df_param.path_current_models,TEAM=training_team,ITER=0))
@@ -478,7 +474,15 @@ if __name__ == '__main__':
 				df_param.l_model_fn.format(\
 					DATADIR=df_param.path_current_models,TEAM=training_team,ITER=iter_i+1))
 
+			if df_param.mice_testing_on: 
+				stats = test_model(df_param,df_param.l_model_fn.format(
+						DATADIR=df_param.path_current_models,TEAM=training_team,ITER=iter_i+1))
+				plotter.plot_test_model(stats)
+				plotter.save_figs('plots/model.pdf')
+				plotter.open_figs('plots/model.pdf')
+				exit()
+
 			if df_param.l_mode == "Mice":
 				increment(df_param)
-	
+
 	print('done!')
